Implement basic methods.
`safe()` - Checks a string to make sure it is only good characters.

`clean()` - Removes bad characters from a string and replaces them with another string.

`replaceFirst()` - Replaces the first instance of a string in another string, if it exists.

`addNew()` - Concatenates a string with another, if it doesn't already exist in the existing string.